adding random_access_compressed_ifstream.hpp/cpp for reading (possibly) compressed mzML, pepXML etc (TPP-world is moving toward treating gzipped XML as a native format).
This is just an istream which chooses its streambuf implementation based on whether the file is gzipped or not.  Could be extended to work with other compression formats, too.

What makes it interesting compared to the classic gzstream implementation is the ability to perform seeks in a reasonably efficient manner.  In the event that a seek is requested (other than a rewind, or tellg()) the file is decompressed once and snapshots of the compression state are made every 1MB or so.  Further seeks are then quite efficient since they don't have to begin at the head of the file.