Add a couple of opt-out bjam commandline switches for building without mz5, or without any binary msdata formats whatsoever.
   --without-mz5           Build without mz5 support

   --without-binary-msdata Build with support for mass spec text formats only
                           (mzML, mzXML, MGF, etc - no binary formats
                           like mz5 or vendor-specifics )

Note that the mz5 build is broken for mingw due to our custom packaging of the hdf5 library - designed for use with autoconf but we use a set of best-guess config headers instead (msvc, darwin, posix - mingw isn't quite posix though).  In my case, this doesn't matter so I'm just making a note of it here.


Former-commit-id: d3072d2de2a906360ffe12ecac82b0f7a65474c5