New AllChromatogramsGraph shows the highest-intensity peaks as they are imported to show import progress and give the user a sneak-peek at their data.
Quick overview of how this works:

As chromatograms are loaded, they are reported to the AllChromatogramsGraph window through a ProgressStatus derived class: ChromatogramLoadingStatus.  This class bins intensity data to compress the large data flow, and notices whenever the average intensity level within a bin falls below a threshold value (calculated as a percentage of the max intensity seen so far).  When this happens, it packages up the "peak" and adds it to a list of finished peaks.  For SRM data, it's a little different, because we get a whole transition at a time, so these peaks are added to the finished peak list immediately.

The finished peak list is periodically retrieved by a background rendering thread.  The background thread does more processing: adding overlapping SRM transitions together, sorting the transitions by maximum intensity, and limiting the total number of peaks to 1000 (otherwise the graph gets too busy and too expensive to render).

The rendering is done by ZedGraph inside AsyncChromatogramsGraph.  This is a subclass of the new user control, AsyncRenderControl, which provides a framework for rendering a bitmap on a background thread.  Another new utility class, Animation, provides support for animating the transitions as they appear in the graph, and also animating the adjustments of the intensity and time range for the graph.  Lots of things can be moving simultaneously, so this class plays a very important role!


Former-commit-id: aef1d0e2ea3a80e74a919f68bc1ef159f2a31756