skyline: merging from my branch for supporting additional protein metadata, branches/work/20140206_skyline_proteinmetadata.
Now we can talk about protein metadata such as accession, preferred_name, gene, and species in addition to existing fields "name" and "description" (the things you can get from a FASTA header).  This metadata is obtained by a combination of parsing name+description and searches on Entrez and UniprotKB.  It will attempt to translate from various protein naming schemes (including IPI and GI numbers) to Uniprot accesion numbers and retrieve gene etc data from the Uniprot servers, possibly visiting Entrez along the way for some translations such as GI numbers.

The protDB schema has been versioned and updated to handle these new fields.  Old unversioned files are considered to be version 0.0, the current version is 0.1 (older versions of Skyline can read v0.1 protDB files, so it's a minor version change, not major)

Several UI changes around this (with more to come) including:
   View Targets menu item is now a choice of items: by Name (as before) or by Accession, PreferredName, or Gene
   Sort Proteins menu item is now a choice of items, sort by Name, Accession, PreferredName, or Gene
   Rename Proteins dialog now has added buttons for conveniently renaming proteins by Accession, PreferredName, or Gene
   Statement completion is aware of the new fields
   Unique Peptides dialog displays gene information when available
   Tooltip text in the sequence tree shows this new information 
   MPP report (and any others that want it) now has gene data
   
The the Skyline document schema has been updated to reflect this change (now at version 1.6), as well as some other recent unrelated changes to the document that weren't in the schema.  I added a schema validation test to Serializable() when called with an SRMDocument, and fixed a couple of unrelated problems in the xsd that this testing revealed, so the schema shouldn't get stale again in the future.

Most of the many changes to existing tests just deal with the fact that we now have another background loader in play, busily updating the protein metadata in the document tree in the background.  This played havoc with tests that assumed the ability to work with a local variable initialized with Skyline.Document, so many WaitForDocumetLoaded() calls had to be added to make sure those doc references don't go stale when the document nodes update with new protein metadata.




Former-commit-id: ae910cc4edfe33b56f603cd200a9a78905cc7e6c