Skyline: adding a test to check whether or not the current test directory has been excluded from virus scanning or not.  When the test directory is being scanned it can lead to file locking problems, and is generally likely to slow down the test cycle.
The test works by creating a little file that, while harmless, contains a standard string used for checking whether an antivirus program is active or not.  If the file disappears after being created, it's assumed that this is due to antivirus quarantine and the following failure message is emitted:

Could not read contents of the (completely harmless!) antivirus test file "eicar_fake_threat.com", probably because it was quarantined by antivirus software.  If your antivirus flagged on "eicar_fake_threat.com", don't panic - that's part of the test (see http://www.eicar.org/86-0-Intended-use.html).  Now go exclude the test directory from further antivirus scrutiny, as it causes file locking problems in the tests.


Former-commit-id: 99a84b6df11d39eee25b78f7d6858537789f75ab