Skyline: avoid treating the potentially narrow m/z range on any one ion mobility separation scan as a SIM indicator by looking at aggregate mz range across drift bins.
I was seeing some strange behavior in Agilent all-ions IMS data, which I finally realized was due to our SIM detection code kicking in.  With IMS data it's possible to have a narrow mz window reported on any one scan, so we have look at the aggregate mz range across all scans with same RT and increasing DT to see if it's really a narrow isolation window setting or not.

Thanks to Nick for code review.


Former-commit-id: 35dc9d70e5ba431f7783bfac5887a5a11fbeb445