Skyline - Multiple Graph Types open at the same time
Now multiple graph types (replicate, peptide, histogram, regression,...) of the same GraphSummary type (Area, RT, MassError) can be
open at the same time. This allows users to look at for example the area replicate comparison and area peptide comparison at the same time without
having to switch back and forth between the two graph types. When a new graph is opened, it will get added as a tab if there is already a graph of the same GraphSummary
type.
This is also fully backward compatible with old view files. (Although from now on the graph types are also stored in the view file)

Other changes:
- changed default dock of library match
- fixed bug in skyline tester that would cause test to fail when clicking "Continue" on the PauseAndContinueForm less than 1 second
after it was shown. (only happened when showing tutorial pages)
- slightly changed several tests to work with these changes


Former-commit-id: b22d412658a9579bed16e58d328ef3e3cd9ab7c6