Skyline: improvements for pasting small molecule transitions lists into Skyline.
In the recent round of small molecule work with Will, I grew frustrated with having to carefully configure the Import Transition List dialog when I could see that he already had recognizable headers in his Excel document. The headers match the column names in the UI, but the set of headers we automatically recognize are different (they have no spaces and are used for settings serialization).

So, I made these improvements to small molecule list paste-to-targets-window (and tool service) behavior:
Recognize the human readable column headers seen in the UI as synonyms for the internal no-spaces strings we already recognize.
Relocate the declaration of known header strings to the model side, removing a dependency on UI in the commandline side
Switch to using the same CSV/TSV format deducer that other transition list readers do


Former-commit-id: 2914dd3a09c98f56725010554132cdbef7304174