Checking in MattC's patch to allow simultaneous use of Waters vendor peak picking and lockmass correction. Formerly both demanded to be first, lockmass correction is now able to wrap Waters native peak picking.
Updated comments in MsDataFileImpl to reflect this.

Also tweaked the ion mobility detection logic: we still can't afford to scan the entire file to see if it has any IMS data, but looking at just the first non-lockmass scan isn't a very robust check, so we now look at the first 20.