Skyline/work/20180427 integrateall (#118)
Make it so that the menu item:
Settings > Integrate All
does not affect calculated areas.

The primary effect of "Integrate All" is now to change the color of the dots next to transitions in the Targets tree. It also affects things like "Peak Found Ratio", but it will not affect things like total area, or group comparisons.