skyline: deal with a race condition in some tutorial tests (TestMs1Tut… (#377)
Skyline: show "None" (localized) instead of a blank line at the head of the list of available IRT Standards. Also, indicate a "no IRT standard" condition in the audit log as "None" instead of "Missing".

The impetus for this work was a race condition in some tutorial tests (TestMs1Tutorial, TestTargetedMSMSTutorial, TestTargetedMSMSTutorialAsSmallMoleculeMasses, TestTargetedMSMSTutorialAsSmallMolecules) that could result in irtStandard being audit logged as either "" or "MISSING" depending on whether certain UI events had time to fire during the test. 

Note that in the code irtStandard.EMPTY was formerly called irtStandard.NULL, but this was against convention.