Fix a problem with commandline "warn on error" mode, where it could f… (#405)
Work around a problem in TeamCity code coverage testing with Skyline commandline "warn on error" mode, due to a race condition between the loader status reporting completion and the document actually finishing all the loading steps. 

In the specific case where we know we're running a test we introduce a brief Thread.Sleep to avoid the race condition. Theoretically a user could still encounter this so I've left a TODO to actually resolve the race condition at some later time.