Prevent the user choosing negative charges for peptide precursors and products on the Transition Settings Filter.
Remove all of the overloads in MessageBoxHelper that take a tab control and page index.
Instead, MessageBoxHelper always calls "EnsureControlTabPagesVisible" which ensures that the correct tab is selected.
When the incorrect tab page is selected, "GetControlMessage" would return the wrong string, since GetControlMessage tries to find a Label that is not hidden.