Skyline: A fairly common support question is "why won't this Skyline … (#502)
Skyline: A fairly common support question is "why won't this Skyline file open?" when in fact the user is actually trying to open a .skyd file or somesuch.  Probably an artifact of Windows hiding file extensions. Now when presented with a .sky<x> file for loading, rather than fail we try to find an identically named  .sky file and load that instead.

If that doesn't work, we now offer a more helpful message, e.g. 'The file you are trying to open ("c:\data\cov_rough.csv") does not appear to be a Skyline document. Skyline documents normally have a ".sky" or ".sky.zip" filename extension and contain XML data.' Formerly we just showed the XML reader exception text, which was not very useful to the user.