Skyline: when trying to paste a seemingly well structure small molecu… (#527)
Skyline: when trying to paste a seemingly well structured small molecule CSV into the Targets window, I was puzzled by a failure that turned out to be due to a trailing space in one of the column headers which prevented us from magically recognizing it. It's better, I think, to just trim() leading and trailing spaces rather than be pedantic about it.