Skyline: when we first started supporting non-peptide data we didn't … (#594)
Skyline: when we first started supporting non-peptide data we didn't have a lot of examples. So we devised a special test mode that can add a non-peptide node to any test document on the fly. We don't really use it any more since we now have actual examples and tests, so now it just serves to confuse new Skyline developers. Thus I'm tearing it out to de-cruft the code base.