Skyline/work/20191206 merge from192 (#773)
Cherry pick from the Skyline 19.2 branch

* Skyline/bugfix192/20191205 prosit library exception (#771)

Fix exception building prosit library when the file cannot be written to.
(Reported on Exception web).

Change it so that only one ProgressStatus object is constructed while building a Prosit Library. A reference to that status object gets passed around to all of the functions that do a segment of work.

Add PrositBuildLibraryTest. This test checks "PrositConfigTest.HasPrositServer". If the production Prosit server is not enabled by the Skyline build then the test does not do anything. Otherwise it tests building a prosit library both with and without an iRT standard.

Make it so that "Requesting predictions from Prosit" status message includes numbers indicating how close to done it is.

Speed up finding distinct peptides. Remove class FuncEqualityComparer-- that class was implementing "GetHashCode" using ToString() on an object whose ToString() was not distinctive-- everything ended up in the same bucket.

Change status update of BlibDb.CreateLibraryFromSpectra so that it does not report 100% completion until after the transaction has been committed.

* Fix ObjectDisposedException in UpgradeManager.update_Complete (#764)
(Reported on exception web)