A couple of changes to test the fix for for the hard-to-reproduce problem with updating MsDataFileUri after we decide that combine_ion_mobility isn't possible for a file.
Just want to run these on Nightly on the test farm, then we'll back them out again.
The idea is to restore some not useful (but normally not harmful) code in SkylineFiles.ms that may provoke the problem if it still exists, and some code in MsDataFilePath that does some console writes to try to hunt down trouble spots.