Skyline/SkylineNightly - added retry logic to re-download data sets o… (#790)
Skyline/SkylineNightly - added optional retry logic to re-download test data sets on test failure, and retry the test if the newly downloaded test data is shown to be different from what was already on disk. This means that the nightly tests no longer have to delete+download existing tutorial data sets, which should improve throughput as well as making us more tolerant of transitory test data server failures.

This behavior is not enabled by default, so that developers are immediately made aware of any issues (and so we don't initiate a big download when you're on an airplane etc).  SkylineNightly tells TestRunner to use it via the  "retrydatadownloads=on" command line argument.