Surequant support in Skyline (#910)
Surequant support

This change adds a checkbox to "Transition Settings > Instrument" which is "Triggered chromatogram extraction"

When that is turned on, Skyline detects where there are gaps in the chromatogram. Skyline constrains peak boundaries so that they do not cross those gaps. Skyline also does not do background acquisition when it is a triggered acquisition. That is, the background area is always zero.