Skyline/work/20200212 surequant peak boundaries (#928)
Fix problems associated with manually adjusted peak boundaries in Triggered Acquisition (SureQuant) data.

If it is Triggered Acquisition then we allow the user to set peak boundaries to any number they want, unconstrained by what the times are in the interpolated chromatograms.

Also fix it so that in GraphChromatogram the feedback that the user sees is always exactly where the peak boundary is going to be placed. Previously the steps would depend on whether you were looking at the Raw or Interpolated chromatogram, even though when you released the mouse the peak boundary would always snap to an interpolated ("evenly spaced") time.