Fix tooltips on Peptide Settings dialog. (#1167)
Fix tooltips on Peptide Settings dialog.
The ToolTip class gets confused if any of the controls with tooltips are no longer part of the form. The ToolTip class figures out what the top level form is by iterating through a Hashtable and taking the first Control.TopLevelControl property. If the first control in the HashTable is no longer attached to the Form, then TopLevelControl ends up being a bogus value, and the tooltips do not work in the dialog.

Add "FormUtil.RemoveTabPage" which removes the tab page, and nulls out all of the tooltips.
Also, change all places in the code which were calling TabPageCollection.RemoveTabPage to now use FormUtil.RemoveTabPage.
Remove now unnecessary code from RefineDlg which was working around this same bug.