Fix Issue 716: When light and heavy precursors have different sets of transitions, all quantifiable transitions should be summed when calculating Light:Heavy ratios  issues
If any of the small molecule transitions under a peptide have not been given a name, then we no longer try to pair up the light and heavy transitions, and instead calculate the ratios by simply summing the transitions that go in the numerator and in the denominator, without making sure that there are the same number of terms in the numerator and denominator.
As a result, we no longer calculate Ratio Dot Product if the any of the transitions do not have a name.