Preserve protein names and/or (for small molecule documents) Molecule List names when exporting .blib spectral libraries.
When .blib files contain this information for small molecules, it is used in "Add to document" to create named Molecule Lists instead of a single "Library Molecules" list as before.

For peptide spectra, if a protein name is present it will be used only if the "Associate Proteins" is not in use, or if no associated document protein is found. When it is used, an peptide group with the protein's name will be created and used instead of the "Library Peptides" group.