Skyline: adding "SkyLint", a simple program that checks for code issues that ReSharper doesn't handle to our satisfaction. In particular, inspections in generated code which ReSharper ignores.
So why isn't this handled with git pre-commit hooks? A few reasons - GitHub doesn't support server-side hooks (as a security issue) and client side there's no agreement on how commit hooks are handled between commandline git and TortoiseGit.

So, this little app is added as a Skyline pre-build step, it throws up a messagebox if it sees a prohibited string in any file matching a file mask, then exits with a non-zero code.

Note that this commit is actually an intentionally broken build, as there are a couple of lingering examples of ".ImageScalingSize = new System.Drawing.Size(" in the code, which this whole project is designed to ferret out.