Total Ion Current normalization (#1224)
Add Total Ion Current as one of the normalization choices.

"Total Ion Current" shows up in the list of normalization choices if either:
a) At least one replicate has a Total Ion Current Area
b) The document has no replicates at all

When normalizing to TIC, we do it in a similar way to how we do median normalization: the raw area is divided by the TIC Area, and then multiplied by the median TIC area across all of the replicates. In this way, the normalized area and the raw area will be close to each other in value.