Skyline/work/20200807 use blib ref spectra proteins (#1180)
Skyline:  when creating .blib spectral libraries, preserve protein names and/or Molecule List names in the .blib RefSpectraProteins table.  When importing small molecule .blib files that contain this information, Library Explorer "Add to Document" will use it to create intelligently named Molecule Lists instead of a single "Library Molecules" list as before.

Requested by Erin and Kaylie.

NOTE: For now, "Add to Document" only uses the RefSpectraProteins table with non-proteomic data to name new PeptideGroup nodes. I've left CONSIDERs to revisit this, but for now there's concern that this would be a potential unwelcome surprise to proteomics users.  When we do eventually decide to enable this for peptides, if a protein name is present it will be used for PeptideGroup naming  only if "Associate Proteins" is not in use, or if no suitable document protein is found by "Associate Proteins". (This logic is in place, just not being used yet.)