Skyline: fix for issue https://skyline.ms/announcements/home/support/thread.view?rowId=49107 in which:   - the user pastes a protein sequence into the Targets tree, it gets the default name "sequence1"   - the user types in a name to replace "sequence1" while background protein metadata search is still underway   - after a few seconds the metadata search completes, and the displayed name mysteriously reverts to "sequence1"
N.B. as this is timing dependent, and our automated tests are mandated to not require internet access, writing a test for this fix (i.e. adding timings to the fake web lookup system) proved to be tricky and finally deemed not worth the effort for a fairly obscure problem.

Reported by Gabriele Philipps