Skyline/work/20201218 non quantitative peak boundaries (#1365)
Ensure that TransitionGroupChromInfo.StartTime and EndTime get set to something even if all of the transitions are non-quantitative.
Skyline tries to set the start/end times to encompass all of the quantitative peaks. If there are no quantitative peaks, then the non-quantitative peaks are used.
(This was something that was causing problems for Panorama).