- fixed inconsistent behavior during DDA search with Close (X) button
Close button was enabled (and would cause errors if search was running) but Cancel button was disabled. However, users are expected to cancel the search first (a separate button), so now the Close button is hidden once the search starts (NB: it does not seem practical to disable it at runtime like the Cancel button; we can only hide it or intercept the Close click and cancel the click). After the search finishes or is canceled (and the search thread terminates properly), then the Close button is made visible again and the main Cancel button is enabled.