Allow specifying more than one crosslink between peptides. (#1428)
Allow specifying more than one crosslink between peptides.

The way that peptides are represented in the Document Tree is now a lot different. Now, a crosslinked peptide is a flat list of one or more peptides, and a flat list of one or more crosslinkers, where each crosslinker has two CrosslinkSite objects in it. A CrosslinkSite is an integer representing which peptide, and an integer representing the amino acid index.

(The old way of representing it was a tree, where nodes were identified by ModificationSite objects, which were modification name and amino acid index. The new way is much simpler)