Prevent the "Peptide"->"Molecule" translation from happening on AlertDlg.
If anyone needs that translation done, it needs to be done before the message is passed to AlertDlg. The AlertDlg has no way of knowing which parts of the message contain user content (filenames, etc) and should therefore never be translated.

(When importing peak boundaries, users sometimes see a message about the column "MoleculeModifiedSequence" being required. This bug was most recently mentioned on a support request from Marek)