Skyline: extend our custom code inspection to detect forgotten uses of PauseTest()… (#1465)
Skyline: extend our custom code inspection to detect forgotten uses of PauseTest(), which can stall our nightly tests. The idea is that any unexpected use of PauseTest() will fail the TeamCity build, so it never makes it to Nightly. And, it's a reminder on your local build to clean that up before a commit. (I recently found myself nearly committing this error on a branch, it's a common enough problem that this seems like a useful check.)

Not reported by any user, internal testing issue only.