Fix problem where value of audit log hash written to XML uses different value of "EnExtraInfo" than what gets written to disk.
The only case where we know this to be a problem is when saving a Skyline document which was originally created while running a unit test. In that case, the value of "EnExtraInfo" that was originally read in will have filenames instead of full paths, but when saving the audit log, the EnExtraInfo will be recalculated from "ExtraInfo", and will have the original full paths again.

In theory, this could also be a problem if any of the localized strings in ColumnCaptions.resx or AuditLogStrings.resx or something have changed from the original Skyline version to now.