Remove .Net 3.5 from system requirements for ClickOnce installer. (#1709)
Remove all of the system requirements from the ClickOnce installer except .Net 4.72.

Also add msvcp140.dll, vcomp140.dll and vcruntime140.dll to pwiz_tools/Skyline/percolator folder so that MSAmanda can run without error on a clean machine.

Also, add the "*.exe.config" files for the executables in the "Method" folder. When the ".exe.config" file is missing, Windows assumes that the exe requires .Net 3.5, and the user gets prompted to install it.