Fix problem where if "Triggered Acquisition" was chosen on Transition Instrument Settings, the reported peak areas would all be 60 times smaller than they should be. (#1740)
If "Triggered chromatogram extraction" is checked on the Transition Instrument Settings, the Area values all end up 60 times smaller than they are supposed to be.

The light:heavy ratios would still be correct, so users might not notice.

(Not reported by a user. I ran into this bug while working on a different feature).