SkylineNightly: improved retry behavior when TeamCity is acting up (#1745)
SkylineNightly: after 30 minutes of failing to download a SkylineTester .lastFinished build, don't give up (as we formerly did) but instead go an additional 90 minutes of retry seeking the .lastSuccessful build instead.

The idea of waiting before failing over to a potentially stale build is to still get some testing in on relatively recent code, while still alerting us to a problem thanks to the lower test count due to that lost half hour.

Also, in case of failure that prevents the tests even starting (e.g. can't download from TC even after retry) just log the problem and exit cleanly rather than throwing. You still get an onscreen representation of the problem (in Notepad, at the tail of the log) but you don't have the problem of a blocking instance of Nightly.exe hanging about and interfering with the next run.