Issue 832: Improve ZedGraph handing of clipboard exceptions (#1751)
Issue 832: Improve ZedGraph handing of clipboard exceptions

The regular "Copy" menu item on the ZedGraph context menu gets replaced with one that we implement in Skyline. Our implementation uses "ClipboardHelper.SetData", which will show the user an error message if the clipboard is locked by a different application.

Added a test "ZedGraphClipboardTest" which makes sure that the menu items put the right type of objects on the clipboard.
That test calls "ClipboardEx.UseInternalClipboard(false)" because the "Copy Metafile" menu item always has to use the system clipboard, because we have never figured out how to create a DataObject with a metafile in it, so it has to use the Win32 function SetClipboardData instead.
Also, the test uses the system clipboard in order to simulate the clipboard being locked by another application.