Fix problems where the "Ratios To" menu item on the Targets Tree did not cooperate with the "Normalize To" menu item on the Peak Area graph.
Now, when you choose a Ratio To on the Targets Tree, that also gets selected on the Peak Area graph.
If you change the "Normalize To" on the Peak Area graph, then that will be reflected on the Targets tree if it is some "Ratio to Label" or "Ratio to Global Standards", but it will be ignored if it's something else.