Improved "Import Peptide Search" wizard for data sets that provide ion mobility information
For a more compact and visually appealing user interface, when IMS values are found in search results Skyline now assumes that ion mobility filtering is desired for chromatogram extraction (formerly there was a checkbox for this), and only asks for resolving power.  If for some reason ion mobility filtering is not desired, Skyline accepts 0 as a valid resolving power meaning "no filtering" (i.e. an infinitely wide window).