Fix for issue 777 "Training IMSDB hangs when raw data files are not present (#1832)
When training an ion mobility database, quite a lot of work is done loading chromatograms before Skyline even tries to access the raw data files to read their ion mobility data.  So when those files turn out to be missing, it's a frustrating user experience. With this change, we check for the raw data before we do anything else. (I didn't reproduce any actual hang, but the delay is indeed frustrating)

Not reported by any user (other than Brendan and Brian)