Use ColumnSelectDlg instead of PasteDlg for importing transition lists (#1845)
Retiring the clunky old transition list import window in favor of the more modern version where the user assigns column types rather than ordering typed columns (in the case of peptides, not even that was possible, so this is a big improvement). Most of the credit here goes to Henry Estberg and Henry Sanford.

N.B. Merging this now in order to meet a UI freeze deadline. I'm still not entirely happy with some of this code - e.g. "comboBoxes[0].SelectedIndex = comboBoxes[1].FindStringExact( xxx );" shows up a lot, that really should be encapsulated. Expect further changes, but not to any RESX files.