Feedback from code review:
Rename new control as LiteDropDownList, and move to Common.Controls namespace.
Clicking on the button when the listbox is showing now closes the listbox.

I think I actually do need to replace the listbox with a menu to get the full ComboBox experience - mouse hover behavior in a listbox isn't quite the same.