Added a new custom code inspection… (#1897)
* Added a new custom code inspection to verify that any english language string that begins with "Error:" is localized such that the L10N version starts with Skyline.Properties.Resources.CommandStatusWriter_WriteLine_Error_ (which is "Error:" in english). I also added this note to https://skyline.ms/wiki/home/development/page.view?name=Skyline%20Custom%20Code%20Inspections :
-------------------------
If you see this:
The zh-CHS language version of resource string <blahblahblah> does not begin with the localized version of "Error:" (see Skyline.Properties.Resources.CommandStatusWriter_WriteLine_Error_)

the problem is that there's an English language string that's probably intended for special handling in pwiz.Skyline.CommandLine.IsErrorMessage(), but the localized version of that string doesn't follow the convention and won't be processed properly at runtime in the case of an error occurring. Probably needs to be sent back to our translation contractors. 
--------------------------