Improved support for mzML files produced by OpenChrom converter
The ProteoWizard mzML reader should not throw when a zlib-compressed array decompresses to 0 length. Zero-length m/z or intensity arrays are unusual but not illegal, and if it really is an error (i.e. the array was declared as being non-zero in length) that still gets flagged by the caller. 

There were additional issues with the OpenChrom converter's mzML output, and we are working with that team to resolve those.  It's unusual to write empty arrays, but not illegal, so that's on us - but there are also mistakes in the mzML file structure that need fixing on their end. They're being very cooperative.

Reported by Skyline user Allison Haase