Added a "Copy to clipboard" button to the PauseForScreenshot dialog as an alternative to Alt-PrtScn when capturing images of Skyline windows for creating and updating tutorials.
When using Alt-PrtScrn it's necessary to first position the window over a clean background to avoid edge pixel schmutz from any intersecting windows. This is extremely tedious when updating several tutorials in multiple languages.

In the "Copy to clipboard" code, after grabbing the window's bitmap we remove any edge-schmutz by determining which color is most commonly seen in the edge pixels and then setting them all to that color. So, no more need to worry about what's overlapped by the window of interest.