Usability improvement to  Edit>Insert>TransitionList
A couple of users have expressed confusion around the new Edit>Insert>TransitionList behavior, wherein you are presented with a big blank text box in which Skyline wants you to paste your transition list. This change attempts to clarify things for the user by moving the prompt from a tiny label above the text box, into the text box itself and with much larger text.

(Developers note: The surprisingly large file count in this commit is just because tests knew a little too much about how the Edit>Insert>TransitionList dialog was implemented. They shouldn't be operating directly on the textbox control to implement the equivalent of a paste action.)