Fix the "Install" class so that it can be used in unit tests. (#1990)
The class was looking at "Assembly.GetExecutingAssembly()" instead of the Skyline executable, which ends up returning something not useful when running as a unit test.
Also, moved the initializing code into a static initializer. It used to be that "Install.Type" would not return the correct value until after the property "Install.Version" was first queried.