Speed up "Find Missing Values" from the Edit Peak Scoring Dialog (#1997)
Sped up generating the list of peptides with missing values from the edit peak scoring dialog.

The code was spending a lot of time in "BookmarkEnumerator.IdentityPath" and "BookmarkEnumerator.ListChromInfos".
Also, added a "ReplicateIndex" property to the "Bookmark" class. That class had been written assuming that a ChromFileInfoId alone is sufficient to identify a particular result file, which is not actually true. The same MsDataFileInfo can appear in multiple replicates so the ReplicateIndex integer is also necessary.