Fixed a problem with reading transition lists defined by m/z only, and with implied isotope labels, e.g. Molecule list name,Molecule name,Precursor m/z,Precursor Charge,Product m/z,Product Charge Compounds,Mol1,430.1,1,236.1,1 Compounds,Mol1,228.1,-1,144.1,-1
Skyline should understand that as a describing a molecule of mass 228.1, with a precursor ion of -1 charge, and a precursor ion of +1 charge and an unidentified isotope label of mass 202. If the "Compounds,Mol1,228.1,-1,144.1,-1" line came first in the list, this was fine, but as is Skyline was reading this as a molecule with mass 430.1 .

Reported by Alex

Developer note - we already had code intended for this situation, but it failed due to the polarity change. It should have been thinking about the calculation of mass from declared m/z and charge instead of only comparing m/z between lines with similar charges.