Prevent ChromInfoList from ever containing null.
Fixed NullReferenceException doing "Apply Peak" when one result file in a replicate was missing results. (Reported anonymously on Exception Web)

ChromInfoList constructor now removes all null's from the list it is passed. There had been obscure scenarios involving multi-file replicates where it was possible to get some of the elements in the list to be null.