Fix and re-enable MidasTest TeamCity failure (#2038)
Fixed error that can happen if the .skyd file is modified while libraries are being loaded.

If TransitionGroupDocNode.UpdateResults notices that the .skyd file has been changed while trying to read the ChromPeak's, the previous TransitionChromInfo's end up being used as if the ChromatogramCache has not been loaded yet. The ChromatogramManager will soon update the SrmDocument with the new data.