Fix an error that could happen when merging .skyd created by different versions of Skyline with different sizes of the ChromPeak data structure. (#2106)
When we use the "FastWrite" API to write out a list of ChromPeaks, the original FileStream would not know to be positioned at the new end of the file. If we then tried to use an ordinary FileStream to write some more data, then FileStream.VerifyOSHandlePosition would throw an error:
The OS handle's position is not what FileStream expected. Do not use a handle simultaneously in one FileStream and in Win32 code or another FileStream. This may cause data loss.
This would only happen if the joining of .skyd files involved where the first .skyd file being joined had been made by the current version of Skyline and the second .skyd file being joined had been created by an older version of Skyline which used a different size for the ChromPeak structure.

(Also, remove unused overload of function "ReadArray" in ChromTransition5)

Co-authored-by: Nicholas Shulman <nicksh@uw.edu>