Fixed a problem with FAIMS on MS2 data
Skyline was not properly filtering for FAIMS in MS2 data, as in Thermo SureQuant, resulting in jagged chromatograms.

Reported by user Arttu

N.B.  While developing this test I discovered that Share Minimized was leaving an empty .imsdb file when it should have had 2 entries due to mishandling of peptide modifications. Fixed that with Nick's help.