Improved detection of lockmass function (hopefully to be obsoleted soon by a new Waters API) (#2283)
Waters lockmass data is typically not wanted in msconvert conversion, or in data processing (e.g. Skyline). Unfortunately the Waters API currently provides no means of identifying the lockmass function, so this is an improvement on the heuristics we use. We now attempt to detect the lockmass function by noting if any function lacks an accompanying chromatogram where the others do not. As part of this work there is a new msconvert option "--ignoreCalibrationScans" which conveniently omits lockmass data from any conversion.