Fix blank library match when spectrum intensities are all zero (#2311)
Fixed bug where Library Match graph disappears if spectrum being displayed has no intensities that are greater than zero (reported by Rich)

The X and Y axes would disappear from the Library Match graph if all of the intensities of the displayed spectrum were zero. This happens a lot with EncyclopeDIA quant libraries where the peptide was detected in the chromatogram library but had no signal in the quantification runs.