Refactor calibration curve code (#2377)
Make it so that "CalibrationCurve" is an abstract class which has subclasses for Linear, Quadratic, LinearInLogSpace, Bilinear
Create a different class "CalibrationCurveMetrics" which is the thing that is displayed in the Document Grid.

Add LinearInLogSpace and Bilinear scenarios to CalibrationScenariosTest. That test generates expected files which are used by the TargetedMS project in LabKey.