Fixes to the way that the background level under the peak is shaded when TransformChrom is something other than Interpolated (#2372)
Fixed the height of the background shading on the chromatogram graph so that is is correct even if a transformation such as Savitzky-Golay or "raw" is being applied. Also changed so that the graph is allowed to extend below the X-axis when transform is "Second Derivative". Removed "First Derivative" transform menu item because it was not hooked up and would always display the same thing as "Interpolated".


Fixes to the way that the background level under the peak is shaded when TransformChrom is something other than Interpolated:
1. Make sure that the shaded background always extends from StartTime to EndTime, even if those times do not correspond to an exact time point in the graph being displayed.
2. Always use the background level from the interpolated chromatogram, nor the transformed chromatogram being displayed.
3. If the TransformChrom is "Second Derivative" then do not fix the Y axis at zero since half of the values in the graph are negative.
4. Remove "First Derivative" menu items from View menu and Chromatogram context menu.